Initial board setup updated
Pieces have now been staggered in order to correctly set up the initial checkers game. Note: Validation not implemented yet. When an opponents player piece is highlighted as a possible move, when the piece is moved to the empty space, the color of the player's piece is lost, however the label '[w]' or '[b]' remains. Pieces that can move can move more than one diagonal space at a time, regardless of whether there is one or more opponent pieces in the way or not.