Move validation (without jumps) added
Moves without jumps are now validated according to the following: 
- Only pieces that are not surrounded by other pieces can be selected to be moved. 
- selected pieces can only be moved one space at a time, diagonally.
- selected pieces can only move forward.
- selected pieces cannot "eat" either their own or other player's pieces next to theirs.
- highlighted hints will only show up for unoccupied, diagonal spaces closest to the selected piece. 

I have also updated the look of the board to mark all unavailable spaces with an [x] (temporary) and changed the colors of the spaces to better resemble a checker board. Player pieces show up on the board as either a white or black circle on the board.